+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Stabilize graph editor test to wait until background jobs are finished before
+  selecting the text fragment.
+
 ## [1.0.0] - 2022-08-23
 
 ### Changed